src/requirements.txt
openai=1.8.0
python-dotenv==1.0.0
fastapi=0.109.0
uvicorn=0.26.0